+++ packages/debug-apps/react-ssr/src/routes/index.tsx
@@ -12,6 +12,7 @@ const components = [
   "dialog",
   "divider",
   "drawer",
+  "file-input",
   "header-bar",
   "icon",
   "icon-button",